+++ src/main/java/com/jnu/festival/domain/booth/service/BoothService.java
@@ -23,7 +23,7 @@ public class BoothService {
 
     //카테고리별 부스 목록 조회
     public List<BoothResponseDTO.BoothListDTO> getBoothList(String location, String period, String category) {
-        // JPA Repository 메서드를 호출하여 필터링된 부스 목록을 가져옴
+
         List<Booth> booths = boothJPARepository.findBooths(location, period, category);
 
         // 결과를 BoothListDTO로 변환
@@ -34,7 +34,7 @@ public class BoothService {
     }
 
     // 부스 상세 조회
-    public BoothResponseDTO.BoothDetailDTO getBoothDetail(int id) {
+    public BoothResponseDTO.BoothDetailDTO getBoothDetail(Long id) {
         Booth booth = boothJPARepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Booth not found with id: " + id));
         return new BoothResponseDTO.BoothDetailDTO(booth);
